backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.16.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1
authlib==1.2.1

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# WebSocket support
websockets==12.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS support
fastapi-cors==0.0.6

# Performance monitoring
psutil==5.9.6
aiohttp==3.9.1
